Moved extra JSONLoader classes to a different file
